cleanup
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@1720 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-moeo/src/fitness/moeoConstraintFitnessAssignment.h
@@ -99,7 +99,7 @@ class moeoConstraintFitnessAssignment : public moeoSingleObjectivization < MOEOT
 			}
 			else{
 				for(unsigned int l=0; l<dim; l++){
-					if (l==to_optimize)
+					if ((int)l==to_optimize)
 						if (_mo.minimizing(l))
 							res-=(normalizer(_mo)[l]) * weight[l];
 						else
